Isolate individual user data
Views including topics, topic, new_topic,
new_entry, edit entry are accessible only when user
is logged in. User will redirected to login page
if not logged in while accessing these pages

Each topic is associated with a user also all the
entries in that topic.

Topics are listed only if the topic owner and
logged in users match. Preventing unauthorised
access with URLs and redirecting it to 404.

Topic owners are automatically added based on the
logged in user.